refactor: change shopping file name
